debug nil naitonality

+++ 14-Wing-Fung-Lau/wk04/5-fri/main.rb
@@ -190,12 +190,14 @@ get '/sanctions/:id/delete' do
     sanction = Sanction.find params[:id]
     sanctions = Sanction.all
     sanction_count = 0
-    sanctions.each do |san|
-        sanction_count += 1 if san.nationality == sanction.nationality
-    end
-    if sanction_count == 1
-        country = Country.find_by(name: sanction.nationality)
-        country.destroy
+    if !sanction.nationality.empty? && !!sanction.nationality.nil?
+        sanctions.each do |san|
+            sanction_count += 1 if san.nationality == sanction.nationality
+        end
+        if sanction_count == 1
+            country = Country.find_by(name: sanction.nationality)
+            country.destroy
+        end
     end
     sanction.destroy
     redirect to('/sanctions')
